+++ gmatpyplus/force_model/__init__.py
@@ -1,3 +1,4 @@
+from .drag_force import *
 from .force_model import *
 from .physical_model import *
 from .point_mass_force import *
